# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI Employee SaaS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /assistants:
    get:
      operationId: listAssistants
      tags: [assistants]
      summary: List all AI assistants for the current user
      security:
        - clerkAuth: []
      responses:
        "200":
          description: List of assistants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Assistant"
    post:
      operationId: createAssistant
      tags: [assistants]
      summary: Create a new AI assistant
      security:
        - clerkAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAssistantBody"
      responses:
        "201":
          description: Created assistant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assistant"
  /assistants/{id}:
    get:
      operationId: getAssistant
      tags: [assistants]
      summary: Get a single assistant
      security:
        - clerkAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Assistant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assistant"
    patch:
      operationId: updateAssistant
      tags: [assistants]
      summary: Update an assistant
      security:
        - clerkAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAssistantBody"
      responses:
        "200":
          description: Updated assistant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assistant"
    delete:
      operationId: deleteAssistant
      tags: [assistants]
      summary: Delete an assistant
      security:
        - clerkAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Knowledge ────────────────────────────────────────────────────────────────
  /assistants/{assistantId}/knowledge:
    get:
      operationId: listKnowledge
      tags: [knowledge]
      summary: List knowledge items for an assistant
      security:
        - clerkAuth: []
      parameters:
        - name: assistantId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of knowledge items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KnowledgeItem"
    post:
      operationId: createKnowledge
      tags: [knowledge]
      summary: Add knowledge to an assistant
      security:
        - clerkAuth: []
      parameters:
        - name: assistantId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateKnowledgeBody"
      responses:
        "201":
          description: Created knowledge item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeItem"
  /knowledge/{id}:
    delete:
      operationId: deleteKnowledge
      tags: [knowledge]
      summary: Delete a knowledge item
      security:
        - clerkAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Chat ─────────────────────────────────────────────────────────────────────
...
  /assistants/{assistantId}/appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      summary: List appointments for an assistant
      security:
        - clerkAuth: []
      parameters:
        - name: assistantId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Book an appointment
      parameters:
        - name: assistantId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAppointmentBody"
      responses:
        "201":
          description: Created appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
  /appointments/{id}:
    patch:
      operationId: updateAppointment
      tags: [appointments]
      summary: Update appointment status
      security:
        - clerkAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAppointmentBody"
      responses:
        "200":
          description: Updated appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"

  # ── Dashboard (wow endpoints) ─────────────────────────────────────────────────
...
  /assistants/{assistantId}/stats:
    get:
      operationId: getAssistantStats
      tags: [dashboard]
      summary: Get statistics for a specific assistant
      security:
        - clerkAuth: []
      parameters:
        - name: assistantId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Assistant statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssistantStats"

  # ── Subscription ──────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    clerkAuth:
      type: http
      scheme: bearer
  schemas:
...
    Assistant:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        name:
          type: string
        businessName:
          type: string
        description:
          type: string
        tone:
          type: string
          enum: [professional, friendly, formal, casual]
        widgetColor:
          type: string
        isActive:
          type: boolean
        totalMessages:
          type: integer
        totalLeads:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, userId, name, businessName, description, tone, widgetColor, isActive, totalMessages, totalLeads, createdAt, updatedAt]
    CreateAssistantBody:
      type: object
      properties:
        name:
          type: string
        businessName:
          type: string
        description:
          type: string
        tone:
          type: string
          enum: [professional, friendly, formal, casual]
        widgetColor:
          type: string
      required: [name, businessName, description]
    UpdateAssistantBody:
      type: object
      properties:
        name:
          type: string
        businessName:
          type: string
        description:
          type: string
        tone:
          type: string
          enum: [professional, friendly, formal, casual]
        widgetColor:
          type: string
        isActive:
          type: boolean
    KnowledgeItem:
      type: object
      properties:
        id:
          type: integer
        assistantId:
          type: integer
        type:
          type: string
          enum: [text, url, pdf]
        title:
          type: string
        content:
          type: string
        sourceUrl:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, assistantId, type, title, content, createdAt]
    CreateKnowledgeBody:
      type: object
      properties:
        type:
          type: string
          enum: [text, url, pdf]
        title:
          type: string
        content:
          type: string
        sourceUrl:
          type: string
      required: [type, title, content]
...
    Appointment:
      type: object
      properties:
        id:
          type: integer
        assistantId:
          type: integer
        conversationId:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        scheduledAt:
          type: string
          format: date-time
        service:
          type: string
        notes:
          type: string
        status:
          type: string
          enum: [pending, confirmed, cancelled, completed]
        createdAt:
          type: string
          format: date-time
      required: [id, assistantId, name, scheduledAt, status, createdAt]
    CreateAppointmentBody:
      type: object
      properties:
        assistantId:
          type: integer
        conversationId:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        scheduledAt:
          type: string
          format: date-time
        service:
          type: string
        notes:
          type: string
      required: [assistantId, name, scheduledAt]
    UpdateAppointmentBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, cancelled, completed]
        scheduledAt:
          type: string
          format: date-time
        notes:
          type: string
...
    AssistantStats:
      type: object
      properties:
        assistantId:
          type: integer
        totalConversations:
          type: integer
        totalMessages:
          type: integer
        totalLeads:
          type: integer
        totalAppointments:
          type: integer
        avgResponseTime:
          type: number
        topQuestions:
          type: array
          items:
            type: string
        leadsByStatus:
          type: object
          additionalProperties:
            type: integer
      required: [assistantId, totalConversations, totalMessages, totalLeads, totalAppointments, avgResponseTime, topQuestions, leadsByStatus]
...